ways-back/requirements.txt
python
django
django_pdb
pygments
djangorestframework
SQLAlchemy
psycopg2
django-cors-headers
social-auth-app-django
django-rest-framework-social-oauth2
django-oauth-toolkit
google-api-python-client